pyproject.toml
[tool.poetry]
name = "nethealth"
version = "0.1.0"
description = "A CLI tool for analyzing network conditions"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.7"
speedtest-cli = "^2.1.3"
requests = "^2.31.0"
psutil = "^5.9.6"
stem = "^1.8.2"
rich = "^13.6.0"

[tool.poetry.dev-dependencies]
black = "^23.10.0"
mypy = "^1.6.1"
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
responses = "^0.23.3"

[tool.poetry.scripts]
nethealth = "nethealth.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true